requirements.txt
paho-mqtt~=1.6.1
selenium~=4.7.2